Bunch of minor renaming.
* src/components/Map => src/components/USMap
* USACountyMap => USMap
* Map => USRiskMap

Mostly this is to better differentiate USMap components (formerly "Map") from
RegionMap and to create space for a USVaccineMap parallel to USRiskMap.
