+++ src/components/HomepageUpsell/HomepageUpsell.tsx
@@ -6,7 +6,7 @@ import {
   Wrapper,
 } from './HomepageUpsell.style';
 import { EventAction, EventCategory } from 'components/Analytics';
-import Map from 'components/Map/Map';
+import USRiskMap from 'components/USMap/USRiskMap';
 import LabelWithChevron from 'components/NewLocationPage/Shared/LabelWithChevron';
 import {
   Experiment,
@@ -35,7 +35,7 @@ const HomepageUpsell: React.FC<{ showHomepageUpsell: boolean }> = ({
                 to="/"
               >
                 <MapWrapper>
-                  <Map showCounties={false} />
+                  <USRiskMap showCounties={false} />
                 </MapWrapper>
                 <div>
                   <LabelWithChevron text="Risk &amp; vaccine tracker" />
