+++ src/components/USMap/USRiskMap.tsx
@@ -1,19 +1,19 @@
 import React from 'react';
 import '../../App.css'; /* optional for styling like the :hover pseudo-class */
-import USACountyMap from './USACountyMap';
+import USMap from './USMap';
 
 interface MapProps {
   showCounties?: boolean;
 }
 
-function Map({ showCounties = false }: MapProps) {
+function USRiskMap({ showCounties = false }: MapProps) {
   return (
     <div className="Map">
       <div className="us-state-map">
-        <USACountyMap showCounties={showCounties} />
+        <USMap showCounties={showCounties} />
       </div>
     </div>
   );
 }
 
-export default Map;
+export default USRiskMap;
