+++ src/screens/HomePage/HomePage.tsx
@@ -1,7 +1,7 @@
 import React, { useRef, useEffect, useState, useMemo } from 'react';
 import Fade from '@material-ui/core/Fade';
 import { useLocation } from 'react-router-dom';
-import Map from 'components/Map/Map';
+import USRiskMap from 'components/USMap/USRiskMap';
 import { NavBarSearch } from 'components/NavBar';
 import { NavAllOtherPages } from 'components/NavBar';
 import AppMetaTags from 'components/AppMetaTags/AppMetaTags';
@@ -137,7 +137,7 @@ export default function HomePage() {
               />
             </ColumnCentered>
 
-            <Map showCounties={showCounties} />
+            <USRiskMap showCounties={showCounties} />
 
             <ColumnCentered $topBottomSpacing={true}>
               <HorizontalThermometer />
